pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
version = "2025.7.1.8"
name = "fwk_common"
dependencies = ["PyYAML","python-dateutil","Jinja2"]
requires-python = ">= 3.8"
authors = [
  {name = "Dan Short", email = "dshort@cossmn.cossmn"}
]
maintainers = [
  {name = "Dan Short", email = "dshort@cossmn.cossmn"}
]
description = "Framework Common function Library"
readme = "README.md"
license = "GPL-3.0-or-later"
keywords= []
[project.urls]
Homepage = "https://github.com/dshortmn/fwk_common"
"Bug Tracker" = "https://github.com/dshortmn/fwk_common/issues"

[tool.setuptools.packages.find]
where = ["."] # Specify where Setuptools should look for packages